export IFilesClient and move to src
