+++ appservice/src/IFilesClient.ts
